WIP refactoring importer
It brings in nodes, materials and textures now but doesn't create meshes yet
